+++ report.go
@@ -6,44 +6,45 @@ package antler
 import (
 	"context"
 	"encoding/gob"
-	"errors"
 	"fmt"
 	"io"
-	"os"
+	"runtime/debug"
 
 	"github.com/heistp/antler/node"
 )
 
-// A reporter can process data items from the node and take some action, such as
-// saving results, generating plots, or emitting logs.
+// A reporter can process data items from the node. It is run in a pipeline,
+// where data items are received on the in channel, and sent on the out channel.
+// Reporters may consume, emit or forward data items. Reporters should forward
+// any unrecognized or unhandled items.
 //
-// When report is called, the implementation must handle the input
-// asynchronously, according to the documentation for reportIn.
+// Reporters may return at any time, with or without an error. Any remaining
+// data on their in channel will be forwarded to the out channel.
 //
-// The report method may be called during TestRun execution multiple times,
-// possibly concurrently, so it must be safe for concurrent use.
+// Reporters may use the given Context to react to cancellation signals, and if
+// canceled, should return the error from context.Cause(ctx). Reporters may also
+// ignore the Context. In any case, they should expect that partial input data
+// is possible, in which case an error should be returned.
 //
-// If a reporter needs to finalize its work, it may also implement io.Closer.
-// If it does, Close will be called after all reports are complete.
+// Both in and out are always non-nil channels.
+//
+// If a reporter is the first stage in a pipeline with no input, the in channel
+// will be closed immediately with no input data.
+//
+// If a reporter is the last stage in a pipeline, it may send nothing to out.
+// For configuration flexibility, most reports should forward data to out as
+// usual, unless they are certain to be the last stage in the pipeline.
+//
+// Reporters may read and write test output using the given 'rwer'.
 type reporter interface {
-	report(reportIn)
+	report(ctx context.Context, in <-chan any, out chan<- any, rw rwer) error
 }
 
-// reportIn is sent to a reporter to do its work.
-//
-// The goroutine that handles the reportIn must receive from the data channel
-// until it's closed. It may send any errors on errc, and must send reportDone
-// on errc when complete.
-//
-// The data channel may be closed by the sender at any time to cancel the
-// report. Therefore, receivers must be prepared to handle partial input, and
-// complete as soon as possible.
-//
-// The embedded writerer provides an Open method for writing result data.
-type reportIn struct {
-	writerer
-	data chan any
-	errc chan error
+// readerer wraps the Reader method, to return a ReadCloser for reading test
+// output. The name parameter identifies the result data according to the
+// underlying implementation, and is typically a filename, or filename suffix.
+type readerer interface {
+	Reader(name string) (io.ReadCloser, error)
 }
 
 // writerer wraps the Writer method, to return a WriteCloser for writing test
@@ -53,17 +54,26 @@ type writerer interface {
 	Writer(name string, overwrite bool) (io.WriteCloser, error)
 }
 
-// reportDone is sent on the error channel by reporters to indicate they are
-// done processing.
-var reportDone = errors.New("report done")
+// rwer groups the readerer and writerer interfaces.
+type rwer interface {
+	readerer
+	writerer
+}
 
-// Report represents the report configuration.
-type Report struct {
-	reporters
+// Report represents a list of reporters.
+type Report []reporters
+
+// report returns an equivalent report instance.
+func (r Report) report() (t report) {
+	for _, p := range r {
+		t = append(t, p.reporter())
+	}
+	return
 }
 
 // reporters is a union of the available reporters.
 type reporters struct {
+	Analyze          *Analyze
 	EmitLog          *EmitLog
 	ChartsFCT        *ChartsFCT
 	ChartsTimeSeries *ChartsTimeSeries
@@ -73,6 +83,8 @@ type reporters struct {
 // reporter returns the only non-nil reporter implementation.
 func (r *reporters) reporter() reporter {
 	switch {
+	case r.Analyze != nil:
+		return r.Analyze
 	case r.EmitLog != nil:
 		return r.EmitLog
 	case r.ChartsFCT != nil:
@@ -86,134 +98,176 @@ func (r *reporters) reporter() reporter {
 	}
 }
 
-// reports is a slice of Report's, with some convenience methods.
-type reports []Report
+// report is a Report list with the reporters unions resolved to implementations
+// of the reporter interface.
+type report []reporter
 
-// reporters returns a slice of reporters for each Report.
-func (s reports) reporters() (reps []reporter) {
-	for _, r := range s {
-		reps = append(reps, r.reporter())
-	}
-	return
+// add appends another report to this one.
+func (r report) add(other report) report {
+	return append(r, other...)
 }
 
-// reporterStack is a stack of reporters used when running a TestRun hierarchy.
-type reporterStack [][]reporter
+// pipeline confines goroutines to run the reporters in a pipeline. See the
+// reporter interface documentation for its contract.
+//
+// If in is not nil, the caller is expected to send to in and close it when
+// done. If out is not nil, the caller is expected to receive all items from it
+// until closed.
+//
+// If the report has no reporters, a nopReport is added so that pipeline's
+// contract is satisfied and it doesn't hang.
+//
+// The returned error channel receives any errors that occur, and is closed when
+// the pipeline is done.
+func (r report) pipeline(ctx context.Context, in chan any, out chan any,
+	rw rwer) <-chan error {
+	if len(r) == 0 {
+		r = append(r, nopReport{})
+	}
+	err := make(chan error)
+	ec := make(chan error)
+	cc := make([]chan any, len(r)+1)
+	if cc[0] = in; cc[0] == nil {
+		cc[0] = make(chan any)
+		close(cc[0])
+	}
+	for i := 1; i < len(r); i++ {
+		cc[i] = make(chan any, dataChanBufLen)
+	}
+	var g int
+	if cc[len(r)] = out; cc[len(r)] == nil {
+		cc[len(r)] = make(chan any, dataChanBufLen)
+		g++
+		go func() {
+			defer func() {
+				ec <- nil
+			}()
+			for range cc[len(r)] {
+			}
+		}()
+	}
+	for i, t := range r {
+		g++
+		go func(t reporter, in <-chan any, out chan<- any) {
+			var e error
+			defer func() {
+				for a := range in {
+					out <- a
+				}
+				close(out)
+				if p := recover(); p != nil {
+					e = fmt.Errorf("pipeline panic in %T: %s\n%s",
+						t, p, string(debug.Stack()))
+				}
+				ec <- e
+			}()
+			e = t.report(ctx, in, out, rw)
+		}(t, cc[i], cc[i+1])
+	}
+	go func() {
+		for i := 0; i < g; i++ {
+			if e := <-ec; e != nil {
+				err <- e
+			}
+		}
+		close(err)
+	}()
+	return err
+}
 
-// push adds a slice of reporters to the stack.
-func (s *reporterStack) push(r []reporter) {
+// tee confines goroutines to pipeline this report to concurrent pipelines for
+// each of the given reports. The output for each "to" report is nil. The
+// returned error channel receives any errors that occur, and is closed when
+// the tee completes.
+func (r report) tee(ctx context.Context, rw rwer, in chan any,
+	to ...report) <-chan error {
+	var c []chan any
+	for range to {
+		c = append(c, make(chan any, dataChanBufLen))
+	}
+	t := tee(c...)
+	var ec []<-chan error
+	ec = append(ec, r.pipeline(ctx, in, t, rw))
+	for i, p := range to {
+		ec = append(ec, p.pipeline(ctx, c[i], nil, rw))
+	}
+	oc := make(chan error)
+	go func() {
+		for e := range mergeErr(ec...) {
+			oc <- e
+		}
+		close(oc)
+	}()
+	return oc
+}
+
+// reportStack is a stack of reports used when running a TestRun hierarchy.
+type reportStack []report
+
+// push adds a report to the stack.
+func (s *reportStack) push(r report) {
 	*s = append(*s, r)
 }
 
-// pop pops a slice of reporters from the stack.
-func (s *reporterStack) pop() {
+// pop pops a report from the stack.
+func (s *reportStack) pop() (r report) {
+	r = (*s)[len(*s)-1]
 	*s = (*s)[:len(*s)-1]
 	return
 }
 
-// list returns a flat list of reporters in the stack.
-func (s *reporterStack) list() (l []reporter) {
+// report returns a flat list of reporters in the stack.
+func (s *reportStack) report() (rr report) {
 	for _, r := range *s {
-		l = append(l, r...)
+		rr = append(rr, r...)
 	}
 	return
 }
 
-// size returns the numbers of reporters in the stack.
-func (s *reporterStack) size() (sz int) {
-	for _, r := range *s {
-		sz += len(r)
-	}
-	return
+// nopReport is a reporter that does nothing.
+type nopReport struct {
 }
 
-// tee receives data from the given channel, and sends it to each reporter in
-// the stack. On the first error, the context is canceled. After data is read
-// in full, the first error, if any, is returned.
-func (s *reporterStack) tee(cancel context.CancelCauseFunc, data chan any,
-	wr writerer) (err error) {
-	ec := make(chan error)
-	var cc []chan any
-	for _, r := range s.list() {
-		c := make(chan any, dataChanBufSize)
-		cc = append(cc, c)
-		r.report(reportIn{wr, c, ec})
-	}
-	n := s.size()
-	a := newAnalysis()
-	dc := data
-	for n > 0 || dc != nil {
-		select {
-		case e := <-ec:
-			if e == reportDone {
-				n--
-				break
-			}
-			if err == nil {
-				err = e
-				cancel(e)
-			}
-		case d, ok := <-dc:
-			if !ok {
-				dc = nil
-				a.analyze()
-				for _, c := range cc {
-					c <- a
-					close(c)
-				}
-				break
-			}
-			a.add(d)
-			if e, ok := d.(error); ok && err == nil {
-				err = e
-			}
-			for _, c := range cc {
-				c <- d
-			}
-		}
-	}
+// report implements reporter
+func (nopReport) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) (err error) {
 	return
 }
 
 // EmitLog is a reporter that emits LogEntry's to files and/or stdout.
+//
+// TODO implement log sorting
 type EmitLog struct {
 	// To lists the destinations to send output to. "-" sends output to stdout,
 	// and everything else sends output to the named file. If To is empty,
 	// output is emitted to stdout.
 	To []string
+
+	// Sort, if true, indicates to gather and sort the logs.
+	Sort bool
 }
 
 // report implements reporter
-func (l *EmitLog) report(in reportIn) {
-	var f simpleReportFunc = l.reportOne
-	f.report(in)
-}
-
-// reportOne runs one EmitLog reporter.
-func (l *EmitLog) reportOne(in reportIn) (err error) {
-	ww := []io.WriteCloser{os.Stdout}
+func (l *EmitLog) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) (err error) {
+	ww := []io.WriteCloser{stdoutWriter{}}
 	defer func() {
 		for _, w := range ww {
-			if w != nil && w != os.Stdout {
-				w.Close()
-			}
+			w.Close()
 		}
 	}()
 	if len(l.To) > 0 {
 		ww = ww[:0]
 		for _, s := range l.To {
-			if s == "-" {
-				ww = append(ww, os.Stdout)
-				continue
-			}
 			var w io.WriteCloser
-			if w, err = in.Writer(s, true); err != nil {
+			if w, err = rw.Writer(s, true); err != nil {
 				return
 			}
 			ww = append(ww, w)
 		}
 	}
-	for d := range in.data {
+	for d := range in {
+		out <- d
 		switch v := d.(type) {
 		case node.LogEntry, node.Error:
 			for _, w := range ww {
@@ -226,18 +280,15 @@ func (l *EmitLog) reportOne(in reportIn) (err error) {
 	return
 }
 
-// SaveFiles is a reporter that saves FileData.
+// SaveFiles is a reporter that saves FileData. If Consume is true, FileData
+// items are not forwarded to the out channel.
 type SaveFiles struct {
+	Consume bool
 }
 
 // report implements reporter
-func (s *SaveFiles) report(in reportIn) {
-	var f simpleReportFunc = s.reportOne
-	f.report(in)
-}
-
-// reportOne runs one SaveFiles reporter.
-func (s *SaveFiles) reportOne(in reportIn) (err error) {
+func (s *SaveFiles) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) (err error) {
 	m := make(map[string]io.WriteCloser)
 	defer func() {
 		for n, w := range m {
@@ -245,15 +296,16 @@ func (s *SaveFiles) reportOne(in reportIn) (err error) {
 			delete(m, n)
 		}
 	}()
-	for d := range in.data {
+	for d := range in {
 		var fd node.FileData
 		var ok bool
 		if fd, ok = d.(node.FileData); !ok {
+			out <- d
 			continue
 		}
 		var w io.WriteCloser
 		if w, ok = m[fd.Name]; !ok {
-			if w, err = in.Writer(fd.Name, true); err != nil {
+			if w, err = rw.Writer(fd.Name, true); err != nil {
 				return
 			}
 			m[fd.Name] = w
@@ -261,55 +313,116 @@ func (s *SaveFiles) reportOne(in reportIn) (err error) {
 		if _, err = w.Write(fd.Data); err != nil {
 			return
 		}
+		if !s.Consume {
+			out <- d
+		}
 	}
 	return
 }
 
-// saveData is a WriteCloser that implements reporter to save result data using
-// gob.
-type saveData struct {
+// expects to be the first stage in a pipeline, so "in" is first discarded.
+//
+// If an decoding error occurs, the error is returned immediately.
+//
+// If the Context is canceled, sending is stopped and the error from
+// context.Cause() is returned.
+type readData struct {
+	io.ReadCloser
+}
+
+// report implements reporter
+func (r readData) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) (err error) {
+	defer r.Close()
+	for range in {
+	}
+	c := gob.NewDecoder(r)
+	for {
+		var a any
+		if err = c.Decode(&a); err != nil {
+			if err == io.EOF {
+				err = nil
+			}
+			return
+		}
+		out <- a
+		select {
+		case <-ctx.Done():
+			err = context.Cause(ctx)
+			return
+		default:
+		}
+	}
+}
+
+// writeData is a WriteCloser and reporter that writes data using gob. writeData
+// expects to be the final stage in a pipeline, so all data is consumed.
+//
+// If an encoding error occurs, the error is returned immediately.
+//
+// If the data includes any errors, the first error is returned after reading
+// and saving all the data.
+type writeData struct {
 	io.WriteCloser
 }
 
 // report implements reporter
-func (s saveData) report(in reportIn) {
-	var f simpleReportFunc = s.reportOne
-	f.report(in)
+func (w writeData) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) error {
+	defer w.Close()
+	c := gob.NewEncoder(w)
+	var f error
+	for d := range in {
+		if e := c.Encode(&d); e != nil {
+			return e
+		}
+		if e, ok := d.(error); ok && f == nil {
+			f = e
+		}
+	}
+	return f
 }
 
+// rangeData is a reporter that sends data from its slice to out. rangeData
+// expects to be the first stage in a pipeline, so "in" is first discarded.
+//
+// If the Context is canceled, sending is stopped and the error from
+// context.Cause() is returned.
+type rangeData []any
+
 // report implements reporter
-func (s saveData) reportOne(in reportIn) (err error) {
-	defer s.Close()
-	c := gob.NewEncoder(s)
-	for d := range in.data {
-		switch d.(type) {
-		case node.FileData, analysis:
-			continue
-		}
-		if err = c.Encode(&d); err != nil {
+func (r rangeData) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) (err error) {
+	for range in {
+	}
+	for _, a := range r {
+		out <- a
+		select {
+		case <-ctx.Done():
+			err = context.Cause(ctx)
 			return
+		default:
 		}
 	}
 	return
 }
 
-// simpleReportFunc is a function that may be called in a separate goroutine for
-// each Test, independently and concurrently. It implements reporter to run a
-// goroutine for each call, and provide boilerplate cleanup.
-type simpleReportFunc func(reportIn) error
+// appendData is a reporter that buffers data in its slice. appendData expects
+// to be the final stage in a pipeline, so all data is consumed.
+//
+// If the data includes any errors, the first error is returned after reading
+// and buffering all the data.
+type appendData []any
 
 // report implements reporter
-func (r simpleReportFunc) report(in reportIn) {
-	go func() {
-		var e error
-		defer func() {
-			if e != nil {
-				in.errc <- e
-			}
-			for range in.data {
-			}
-			in.errc <- reportDone
-		}()
-		e = r(in)
-	}()
+func (a *appendData) report(ctx context.Context, in <-chan any, out chan<- any,
+	rw rwer) error {
+	var f error
+	for d := range in {
+		*a = append(*a, d)
+		if e, ok := d.(error); ok && f == nil {
+			f = e
+		}
+	}
+	return f
 }